% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icc.R
\name{Tbll_icc}
\alias{Tbll_icc}
\alias{Tbll_icc.matrix}
\alias{Tbll_icc.data.frame}
\title{Interklass Korrelationskoeffizient}
\usage{
Tbll_icc(x, ...)

\method{Tbll_icc}{matrix}(
  x,
  ...,
  missing = TRUE,
  alpha = 0.05,
  lmer = TRUE,
  check.keys = FALSE,
  type = c(1, 4)
)

\method{Tbll_icc}{data.frame}(
  data,
  ...,
  rater = NULL,
  obs = NULL,
  value = NULL,
  type = 2 + 3,
  missing = TRUE,
  alpha = 0.05,
  lmer = TRUE,
  check.keys = FALSE
)
}
\arguments{
\item{x, ...}{an prepare_data2}

\item{...}{an psych::ICC}

\item{missing, alpha, lmer, check.keys}{an psych::ICC}

\item{type}{Number 1 bis 6}

\item{data}{data.frame}

\item{rater}{Messinatrument oder Rater}

\item{obs}{eindeutige Id kann auch aus meheren Parametern bestehen}

\item{value}{Messwert}
}
\value{
data.frame   tibble: x × 8
 type  ICC   lower bound upper bound F       df1   df2 p
}
\description{
intraclass correlation coefficient (ICC)
}
\details{
Der Intraklass-Korrelationskoeffizient (ICC) beurteilt die Stärke und Richtung der
Korrelation (Zusammenhang) zwischen den wiederholten Messungen eines Merkmals am selben Individuum.
}
\examples{
#'
#'# Less than 0.50: Poor reliability
# Between 0.5 and 0.75: Moderate reliability
# Between 0.75 and 0.9: Good reliability
# Greater than 0.9: Excellent reliability

#Suppose four different judges were asked to rate the quality of 10 different college entrance exams.
# The results are shown below:
data <- data.frame(
  A = c(1, 1, 3, 6, 6, 7, 8, 9, 8, 7),
  B = c(2, 3, 8, 4, 5, 5, 7, 9, 8, 8),
  C = c(0, 4, 1, 5, 5, 6, 6, 9, 8, 8),
  D = c(1, 2, 3, 3, 6, 4, 6, 8, 8, 9)
)
data

Tbll_icc(data)

# The intraclass correlation coefficient (ICC) turns out to be 0.782.
#
# Based on the rules of thumb for interpreting ICC,
# we would conclude that an ICC of 0.782 indicates that the exams
# can be rated with “good” reliability by different raters.



DF <- stp25tools::get_data("
nr   ratings obs    rate
1        9   obj1  rater1
2        6   obj2  rater1
3        8   obj3  rater1
4        7   obj4  rater1
5       10   obj5  rater1
6        6   obj6  rater1
7        2   obj1  rater2
8        1   obj2  rater2
9        4   obj3  rater2
10       1   obj4  rater2
11       5   obj5  rater2
12       2   obj6  rater2
13       5   obj1  rater3
14       3   obj2  rater3
15       6   obj3  rater3
16       2   obj4  rater3
17       6   obj5  rater3
18       4   obj6  rater3
19       8   obj1  rater4
20       2   obj2  rater4
21       8   obj3  rater4
22       6   obj4  rater4
23       9   obj5  rater4
24       7   obj6  rater4")

Tbll_icc(DF,
     value = "ratings",
     obs = "obs",
     rater = "rate")
}
