% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kapa.R
\name{Tbll_kappa}
\alias{Tbll_kappa}
\title{Cohens Kappa fuer k Urteilsauspraegungen}
\usage{
Tbll_kappa(
  x,
  ...,
  include.ci = TRUE,
  ci.level = 0.95,
  include.weighted = TRUE,
  include.unweighted = TRUE,
  weights = "Equal-Spacing"
)
}
\arguments{
\item{x}{Objekt Kappa aus VCD}

\item{include.weighted, include.unweighted}{kappa statistic}

\item{weights}{an vcd::Kappa   c("Equal-Spacing", "Fleiss-Cohen")}
}
\description{
Kategorisierung von Cohens-Kappa-Werten
Wert von Kappa Ausmass der Uebereinstimmung
< 0.20  =  nicht ausreichend (poor)
0.21 bis 0.40  =  hinreichend (fair)
0.41 bis 0.60  =  moderat (moderat)
0.61 bis 0.80  =  gut (good)
0.81 bis 1.0  =  sehr gut (very good)
}
\examples{

Botulinum <- data.frame(
A= factor(c(rep(1, 14), rep(1, 3),
            rep(0, 5),rep(0, 18)),
          1:0, c("+", "-")),
B= factor(c(rep(1, 14), rep(0, 3),
            rep(1, 5),rep(0, 18)),
          1:0, c("+", "-")))


vcd::Kappa(xtabs(~A+B, Botulinum))
Tbll_kappa(xtabs(~A+B, Botulinum))

Tbll_kappa(~A+B, Botulinum)



}
